requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
APScheduler==3.6.3
async-timeout==4.0.1
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
cryptography==36.0.0
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
frozenlist==1.2.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hexbytes==0.2.2
hpack==4.0.0
httpcore==0.14.3
httpx==0.21.1
hyperframe==6.0.1
idna==2.10
importlib-metadata==4.8.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.7.19
parsimonious==0.8.1
pip==21.3.1
protobuf==3.19.1
pycparser==2.21
pycryptodome==3.12.0
pyOpenSSL==19.1.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dotenv==0.19.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.24.0
rfc3986==1.5.0
rlp==1.2.0
setuptools==59.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
toolz==0.11.2
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.25.11
varint==1.0.2
web3==5.25.0
websockets==9.1
Werkzeug==1.0.1
wheel==0.37.0
WTForms==2.3.3
yarl==1.7.2
zipp==3.6.0
